server/requirements.txt
requests>=2.31.0
crewai==0.201.1
crewai-tools==0.75.0
googlesearch_python==1.3.0
langchain_community==0.3.29
langchain_groq==0.3.7
python-dotenv==1.1.1
gunicorn==23.0.0
whitenoise==6.9.0
django-cors-headers==4.7.0
openai==1.109.1
channels
GitPython
semgrep
langchain-openai